GOS 2025/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.12.1
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
python-socketio>=5.10.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
geoalchemy2>=0.14.2
psycopg2-binary>=2.9.9
shapely>=2.0.0

